seem to have more or less cleaned up all the refactor noise

+++ engine/entity_class.go
@@ -8,5 +8,5 @@ package engine
 
 type EntityClass interface {
 	Name() string
-	DefaultSpawnRequest(position [2]int16) EntitySpawnRequest
+	DefaultSpawnRequest(position [2]int32) EntitySpawnRequest
 }
